Update utility.inc.php
fixing deprecated implicit cast

+++ server/inc/utility.inc.php
@@ -31,8 +31,8 @@ function timeduration($time,$inputunit="hours"){
 	$ms=$time-floor($time);
 	$printms=sprintf("%03d",floor($ms*1000));
 	$s=$time % 60;
-    $m=(($time-$s) / 60) % 60;
-    $h=floor($time / 3600);
+    $m=(($time-(float)$s) / 60) % 60;
+    $h=floor($time / (float)3600);
 
 	$output=$h.":".substr("0".$m,-2).":".substr("0".$s,-2);
 	if($time<1 and $time<>0){
